chore: 파일 디렉토리 변경
#151
