+++ src/main/java/com/barter/domain/trade/immediatetrade/dto/request/CreateImmediateTradeReqDto.java
@@ -1,4 +1,4 @@
-package com.barter.domain.trade.immediatetrade.service.dto.request;
+package com.barter.domain.trade.immediatetrade.dto.request;
 
 import com.barter.domain.product.entity.RegisteredProduct;
