+++ src/main/java/com/barter/domain/trade/immediatetrade/dto/request/UpdateImmediateTradeReqDto.java
@@ -1,4 +1,4 @@
-package com.barter.domain.trade.immediatetrade.service.dto.request;
+package com.barter.domain.trade.immediatetrade.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
