+++ src/main/java/com/barter/domain/trade/immediatetrade/repository/ImmediateTradeRepository.java
@@ -1,10 +1,10 @@
-package com.barter.domain.trade.immediatetrade.service.repository;
+package com.barter.domain.trade.immediatetrade.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.barter.domain.trade.immediatetrade.service.entity.ImmediateTrade;
+import com.barter.domain.trade.immediatetrade.entity.ImmediateTrade;
 
 public interface ImmediateTradeRepository extends JpaRepository<ImmediateTrade, Long> {
