+++ src/main/java/com/barter/domain/trade/immediatetrade/service/ImmediateTradeService.java
@@ -1,4 +1,4 @@
-package com.barter.domain.trade.immediatetrade.service.service;
+package com.barter.domain.trade.immediatetrade.service;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,13 +18,13 @@ import com.barter.domain.product.repository.RegisteredProductRepository;
 import com.barter.domain.product.repository.SuggestedProductRepository;
 import com.barter.domain.product.repository.TradeProductRepository;
 import com.barter.domain.trade.enums.TradeStatus;
-import com.barter.domain.trade.immediatetrade.service.dto.request.CreateImmediateTradeReqDto;
-import com.barter.domain.trade.immediatetrade.service.dto.request.CreateTradeSuggestProductReqDto;
-import com.barter.domain.trade.immediatetrade.service.dto.request.UpdateImmediateTradeReqDto;
-import com.barter.domain.trade.immediatetrade.service.dto.request.UpdateStatusReqDto;
-import com.barter.domain.trade.immediatetrade.service.dto.response.FindImmediateTradeResDto;
-import com.barter.domain.trade.immediatetrade.service.entity.ImmediateTrade;
-import com.barter.domain.trade.immediatetrade.service.repository.ImmediateTradeRepository;
+import com.barter.domain.trade.immediatetrade.dto.request.CreateImmediateTradeReqDto;
+import com.barter.domain.trade.immediatetrade.dto.request.CreateTradeSuggestProductReqDto;
+import com.barter.domain.trade.immediatetrade.dto.request.UpdateImmediateTradeReqDto;
+import com.barter.domain.trade.immediatetrade.dto.request.UpdateStatusReqDto;
+import com.barter.domain.trade.immediatetrade.dto.response.FindImmediateTradeResDto;
+import com.barter.domain.trade.immediatetrade.entity.ImmediateTrade;
+import com.barter.domain.trade.immediatetrade.repository.ImmediateTradeRepository;
 
 import lombok.RequiredArgsConstructor;
